server/requirements.txt
anyio==4.8.0
asgiref==3.8.1
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
Django==4.2
django-cors-headers==4.7.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.4
httpx==0.25.0
idna==3.10
mongoengine==0.29.1
motor==3.3.2
PyJWT==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
